+++ modulo.cpp
@@ -7,10 +7,10 @@
 /// Parent:
 template <unsigned m> struct Mint {
     Mint():val(0){}
-    Mint(int raw):val((raw%m+m)%m){}
-    Mint(long long raw):val((raw%m+m)%m){}
+    Mint(int raw):val((raw%(int)m+(int)m)%(int)m){}
+    Mint(long long raw):val((raw%(long long)m+(long long)m)%(long long)m){}
     Mint(unsigned raw):val(raw%m){}
-    Mint(unsigned long long raw):val(raw%m){}
+    Mint(unsigned long long raw):val(raw%(unsigned long long)m){}
     friend Mint operator+(Mint a, Mint b) { return a.val + b.val; }
     friend Mint operator-(Mint a, Mint b) { return m + a.val - b.val; }
     friend Mint operator*(Mint a, Mint b) { return 1ull * a.val * b.val; }
